fix: make detect-hifiberry run only once per host
Don't try to update the /boot/config.txt by rerunning the detect-hifiberry script (from upstream/HiFiBerry OS) since, once a HiFiBerry device was configured, the i2cget calls to detect the device return different values than initially, making subsequent invocations of the detect-hifiberry script fail to detect the device properly and consequentially break the configuration (that may have worked previously).

+++ entrypoint.sh
@@ -1,63 +1,76 @@
 #!/bin/sh
 
+IGNORE_MARKER_FILE=/host/boot/hifiberry-autoconfig.ignore
+
 set -eu
+
+isHostIgnorable() {
+	[ -f $IGNORE_MARKER_FILE ]
+}
+
+markHostAsIgnorable() {
+	# Let the script ignore subsequent invokations
+	touch $IGNORE_MARKER_FILE
+	sync
+}
 
 alsaDeviceDetected() {
 	aplay -l | grep hifiberry | grep -q pcm5102
 }
 
 backupBootConfig() {
-	if [ ! -f /host/boot/config.txt.base ]; then
-		echo 'Creating /boot/config.txt backup at /boot/config.txt.base'
-		cp -f /host/boot/config.txt /host/boot/config.txt.base.tmp
-		mv /host/boot/config.txt.base.tmp /host/boot/config.txt.base
+	BACKUP_FILE=/boot/config.txt.hifiberry-autoconf.bak
+	if [ ! -f /host${BACKUP_FILE} ]; then
+		echo "INFO: Creating /boot/config.txt backup at ${BACKUP_FILE}"
+		cp -f /host/boot/config.txt /host${BACKUP_FILE}.tmp
+		mv /host${BACKUP_FILE}.tmp /host${BACKUP_FILE}
+		sync
+	else
+		echo "INFO: Skipping /boot/config.txt backup since it already exists at ${BACKUP_FILE}"
 	fi
-	cp -f /host/boot/config.txt.base /boot/config.txt.tmp
-	mv /boot/config.txt.tmp /boot/config.txt
-	sync
 }
 
-writeBootConfig() {
-	# Updating /boot/config.txt
+writeHostBootConfig() {
+	echo 'INFO: Writing /boot/config.txt to host'
 	cp -f /boot/config.txt /host/boot/config.txt.tmp
 	mv /host/boot/config.txt.tmp /host/boot/config.txt
 	sync
 }
 
-rebootIfBootConfigChanged() {
-	if [ "$REBOOT_ON_CHANGE" = true ]; then
-		STATUS=0
-		diff /boot/config.txt /host/boot/config.txt || STATUS=$?
-		if [ $STATUS -eq 1 ]; then
-			writeBootConfig
-			echo 'Rebooting the system since /boot/config.txt changed'
-			sleep 3
-			kill -2 1
-		fi
-	else
-		writeBootConfig
-	fi
-}
-
-writeHostAlsaConfigIfNotExist() {
-	if [ -d /host/etc ]; then
-		if [ ! -f /host/etc/asound.conf ]; then
-			echo 'Installing /etc/asound.conf on the host'
-			cp -f /etc/asound.conf /host/etc/asound.conf.tmp &&
-			mv /host/etc/asound.conf.tmp /host/etc/asound.conf &&
-			sync
-		fi
+writeHostALSAConfigIfNotExist() {
+	if [ -d /host/etc ] && [ ! -f /host/etc/asound.conf ]; then
+		echo 'INFO: Installing /etc/asound.conf on the host'
+		cp -f /etc/asound.conf /host/etc/asound.conf.tmp &&
+		mv /host/etc/asound.conf.tmp /host/etc/asound.conf &&
+		sync
 	fi
 }
 
 waitForI2CDevicesToBecomeAvailable() {
-	for _ in $(seq 1 15); do
+	I2S_INIT_TIMEOUT=30
+	for _ in $(seq 1 $I2S_INIT_TIMEOUT); do
 		sleep 1
 		find /dev -path '/dev/i2c*' | grep -q . && return 0 || true
-		echo 'Waiting for I2C device files to become available' >&2
+		echo 'INFO: Waiting for I2C device files to become available' >&2
 	done
-	echo 'Timed out waiting 15s for device files to become available' >&2
+	echo "ERROR: Timed out waiting ${I2S_INIT_TIMEOUT}s for device files to become available" >&2
 	return 1
+}
+
+printDeviceInfo() {
+	echo "INFO: Found I2C devices: $(find /dev -path '/dev/i2c*' | xargs)"
+	echo 'INFO: Found ALSA devices:'
+	aplay -l | grep -E '^card ' | sed -E 's/^/INFO:   /'
+}
+
+rebootMaybe() {
+	if [ "$REBOOT_ON_CHANGE" = true ]; then
+		echo 'INFO: Rebooting the system'
+		sleep 3
+		kill -2 1
+	else
+		echo 'INFO: Skipping reboot since REBOOT_ON_CHANGE=false'
+	fi
 }
 
 # ARGS: EXITCODE MESSAGE
@@ -70,20 +83,25 @@ terminate() {
 	fi
 }
 
-
-if alsaDeviceDetected; then
+if isHostIgnorable; then
+	printDeviceInfo &&
 	echo ready > /tmp/ready &&
-	terminate 0 'Skipping auto-configuration since ALSA device found'
+	terminate 0 "INFO: Skipping HiFiBerry auto-configuration since file $IGNORE_MARKER_FILE exists"
+elif alsaDeviceDetected; then
+	printDeviceInfo &&
+	echo ready > /tmp/ready &&
+	terminate 0 'INFO: Skipping HiFiBerry auto-configuration since ALSA device found'
 else
-	echo 'Detecting HiFiBerry I2C device ...'
+	echo 'INFO: Detecting HiFiBerry I2C device ...'
 	modprobe i2c-dev &&
 	waitForI2CDevicesToBecomeAvailable &&
-	echo "Found I2C devices: $(find /dev -path '/dev/i2c*' | xargs)" &&
-	(echo "DAC+ detection: `i2cget -y 1 0x4d 40`" || true) &&
-	backupBootConfig && # Make detect-hifiberry idempotent, allow restoring config.
+	printDeviceInfo &&
+	backupBootConfig &&
 	detect-hifiberry && # Writes /boot/config.txt eventually, configuring dtoverlay.
-	writeHostAlsaConfigIfNotExist &&
-	rebootIfBootConfigChanged &&
+	writeHostALSAConfigIfNotExist &&
+	writeHostBootConfig &&
+	markHostAsIgnorable &&
+	rebootMaybe &&
 	echo ready > /tmp/ready &&
-	terminate 0 'Configuration is ready' || terminate 1 'ERROR: HiFiBerry auto-configuration failed' >&2
+	terminate 0 'INFO: Configuration is ready' || terminate 1 'ERROR: HiFiBerry auto-configuration failed' >&2
 fi
